Add seperate entry point for node
